Rotation4
RNA seq analysis using DeSeq2. Identified predominant transcripts for DDX6, a gene I was studying in relation to SMARCB1.

I created a for loop to plot expression of genes in a specific file. Here I created dot plots of the expression of genes involved in RNA degradation. While seemingly straight forward, what proved trivial in this for loop was that the axis titles were properly represented but there was no information on the plots. I successfully managed to represent each gene in this pathway.